requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
datetime
jinja2
git+https://github.com/Peter-J/aioairctrl.git
